feat(revision): implement quantity calculation
Quantity.qty_sect/qty_bar change from Integer to Numeric(10,2) - a plain
float risks binary rounding error for BOQ figures. Every Object row, of
any change_status, gets exactly one Quantity row (D9 full snapshot),
created by a shared _create_quantity() helper called from both
determine_change_status and _duplicate_object.

qty_sect: cross-section area x extent (rectangular width*depth or
circular pi*r^2) for axis-point types, or thickness x the polygon's true
3D planar area (Newell's method, not a 2D shoelace formula that would
drop Z and mis-measure a sloped wall) for boundary-point types.
Represents weight instead of volume when the object's material is
steel. qty_bar: total reinforcement weight summed across every
Reinforcement row, 0 for zero rows (valid for steel, VR-11) rather than
an error.

Also fixes QuantityDetail's response schema to use Decimal instead of
float, matching the same rounding-precision rationale, after mypy/
pyright flagged the type mismatch once the model column type changed.

+++ schemas.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from typing import Literal, Self
 
 from pydantic import BaseModel, ConfigDict, Field, model_validator
@@ -239,8 +240,10 @@ class ReinforcementDetail(BaseModel):
 
 
 class QuantityDetail(BaseModel):
-    qty_sect: float
-    qty_bar: float
+    # Decimal, not float - matches Quantity.qty_sect/qty_bar (D23): a plain
+    # float risks binary rounding error for BOQ figures.
+    qty_sect: Decimal
+    qty_bar: Decimal
 
 
 class ObjectDetail(BaseModel):
